Basic updates to allow for compiling on Windows

+++ Source/Player/CSceneTransitionPlayerCoverUncover.cpp
@@ -47,7 +47,7 @@ CSceneTransitionPlayerCoverUncover::CSceneTransitionPlayerCoverUncover(CScenePla
 	mInternals =
 			new CSceneTransitionPlayerCoverUncoverInternals(info.contains(mInfoIsAutoKey), isCover,
 					getSceneTransitionDirection(info.getString(mInfoDirectionKey)),
-					info.getFloat32(mInfoDurationKey, 0.3), sceneTransitionPlayerProcsInfo);
+					info.getFloat32(mInfoDurationKey, 0.3f), sceneTransitionPlayerProcsInfo);
 }
 
 //----------------------------------------------------------------------------------------------------------------------
@@ -74,7 +74,7 @@ void CSceneTransitionPlayerCoverUncover::update(UniversalTimeInterval deltaTimeI
 			case kSceneTransitionDirectionDown:
 				// Up or Down
 				mInternals->mCurrentOffset +=
-						deltaTimeInterval / mInternals->mDurationTimeInterval * viewportPixelSize.mHeight;
+						(Float32) (deltaTimeInterval / mInternals->mDurationTimeInterval) * viewportPixelSize.mHeight;
 				if (mInternals->mCurrentOffset > viewportPixelSize.mHeight) {
 					// Beyond bounds, done
 					mInternals->mCurrentOffset = viewportPixelSize.mHeight;
@@ -86,7 +86,7 @@ void CSceneTransitionPlayerCoverUncover::update(UniversalTimeInterval deltaTimeI
 			case kSceneTransitionDirectionRight:
 				// Left or Right
 				mInternals->mCurrentOffset +=
-						deltaTimeInterval / mInternals->mDurationTimeInterval * viewportPixelSize.mWidth;
+						(Float32) (deltaTimeInterval / mInternals->mDurationTimeInterval) * viewportPixelSize.mWidth;
 				if (mInternals->mCurrentOffset > viewportPixelSize.mWidth) {
 					// Beyond bounds, done
 					mInternals->mCurrentOffset = viewportPixelSize.mWidth;
